parallelized the consolidation more
