+++ phold/.gitignore
@@ -1,2 +1,4 @@
 *.out
 *dir
+rank*
+*.csv
