+++ phold/dispatch.sh
@@ -4,16 +4,17 @@ scriptDir="$(dirname "$(scontrol show job "$SLURM_JOB_ID" | awk -F= '/Command=/{
 echo "$scriptDir"
 
 nodeCount=$1
-threadCount=$2
-width=$3
-height=$4
-eventDensity=$5
-ringSize=$6
-timeToRun=$7
-smallPayload=$8
-largePayload=$9
-largeEventFraction=$10
-prefix=${11}
+ranksPerNode=$2
+threadCount=$3
+width=$4
+height=$5
+eventDensity=$6
+ringSize=$7
+timeToRun=$8
+smallPayload=$9
+largePayload=${10}
+largeEventFraction=${11}
+prefix=${12}
 
 tmpFile=${prefix}.tmp
 timeFile=${prefix}.time
@@ -25,7 +26,7 @@ simFlags="--N $height --M $width --eventDensity $eventDensity --timeToRun ${time
 
 sstFlags="--num-threads $threadCount --print-timing-info=true --parallel-load=SINGLE ${scriptDir}/phold_dist.py"
 
-srunPortion="srun -N $nodeCount --cpus-per-task=$threadCount --ntasks-per-node=1" 
+srunPortion="srun -N $nodeCount --cpus-per-task=$threadCount --ntasks-per-node=$ranksPerNode" 
 
 
 mkdir $outDir
